requirements.txt
Flask==3.0.2
numpy==1.21.5
Pillow==10.2.0
Requests==2.31.0
